+++ CHANGELOG.md
@@ -6,6 +6,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Implements `fmt::Debug` for `Basis2`, `Basis3`, and `AffineMatrix3`
+
+### Changed
+
+- Improves the `fmt::Debug` impls for `Vector`, `Matrix`, `Point`, `Decomposed`,
+  `Quaternion` and `Angle` to make them easier to derive, and have clearer
+  formatting.
+
 ## [v0.7.0] - 2015-12-23
 
 ### Added